try some possible bugfixes for the sigsev

+++ src/smt/smt_context.h
@@ -139,7 +139,7 @@ namespace smt {
         scoped_ptr<base_dependent_expr_state> m_fmls;
 
         svector<double> m_lit_scores[2];
-        svector<double> m_birthdate;
+        svector<unsigned> m_birthdate;
 
 
         // -----------------------------------
